+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 A persistence layer for Hanami
 
 ## v1.0.0.beta2 - 2017-03-02
+### Added
+- [Semyon Pupkov] Allow to define Postgres connection URL as `"postgresql:///mydb?host=localhost&port=6433&user=postgres&password=testpasswd"`
+
 ### Fixed
 - [Marcello Rocha] Fixed migrations MySQL detection of username and password
 - [Luca Guidi] Fixed migrations creation/drop of a MySQL database with a dash in the name